client_agent_example/requirements.txt
requests
python-dotenv
schedule # Для периодических задач, если нужно